% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deploy_learnr_pkg.R
\name{deploy_learnr_pkg}
\alias{deploy_learnr_pkg}
\title{Install learnr tutorials}
\usage{
deploy_learnr_pkg(tutorials = NULL, pkg = NULL, lib = NULL)
}
\arguments{
\item{tutorials}{List of tutorials to update. Default: all tutorials in the tutorials directory}

\item{pkg}{Tutorial package. Default: current directory}

\item{lib}{Library to install pkg. Default: library 1 (the renv library if activated)}
}
\description{
Run this function from the base of the package project directory, with renv active (or set the arguments). Will perform the following actions:

1. git pull

2. update the renv library

3. install \code{pkg} in \code{lib}.

4. regenerate the HTML files for the listed \code{tutorials}

5. make the renv library accessible to all users with \code{chmod -R 755}

6. make the path up to the renv library accessible with \code{chmod 755}
}
\note{
Uses system calls which may not be platform independent.

If you get an \code{error in .validate_signal_args}, try restarting the R session and run again.
}
\examples{
learnr.dashboard:::setup_learnr_pkg("fg3")

learnr.dashboard:::setup_learnr_pkg(c("fg1", "fg2"))
}
