% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learnr_setup.R
\name{.learnr_setup}
\alias{.learnr_setup}
\title{Assert that a renv library is available.
If load=TRUE, set the renv library and system libraries as only libraries.
If load=FALSE, return the renv library and system libraries.
Uses the renv system library if available, else use the default system libraries.}
\usage{
.learnr_setup(renv_lib = NULL, system_libs = NULL, load = TRUE)
}
\description{
Assert that a renv library is available.
If load=TRUE, set the renv library and system libraries as only libraries.
If load=FALSE, return the renv library and system libraries.
Uses the renv system library if available, else use the default system libraries.
}
