% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start_as_tutorial.R
\name{.open_browser}
\alias{.open_browser}
\title{start a browser when the app has loaded}
\usage{
.open_browser(name, logfile, address = "http://127.0.0.1")
}
\description{
start a browser when the app has loaded
}
