% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start_as_tutorial.R
\name{.run}
\alias{.run}
\title{re-implementation of rmarkdown::run.
this version calls learnr.dashboard:::.runApp instead of shiny::runApp}
\usage{
.run(
  file = "index.Rmd",
  dir = dirname(file),
  default_file = NULL,
  auto_reload = TRUE,
  shiny_args = NULL,
  render_args = NULL
)
}
\description{
re-implementation of rmarkdown::run.
this version calls learnr.dashboard:::.runApp instead of shiny::runApp
}
