% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start_as_tutorial.R
\name{.run_tutorial}
\alias{.run_tutorial}
\title{re-implementation of learnr::run_tutorial.
~~if the tutorial html file exists, learnr does not need to write anything.~~
~~this version skips the write permission check if not required.~~
this version just removes the message.}
\usage{
.run_tutorial(name = NULL, package = NULL, shiny_args = NULL)
}
\description{
re-implementation of learnr::run_tutorial.
~~if the tutorial html file exists, learnr does not need to write anything.~~
~~this version skips the write permission check if not required.~~
this version just removes the message.
}
