% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start_as_tutorial.R
\name{.runApp}
\alias{.runApp}
\title{re-implementation of shiny::runApp
this version uses a more restricted range of ports}
\usage{
.runApp(
  appDir = getwd(),
  port = getOption("shiny.port"),
  launch.browser = getOption("shiny.launch.browser", interactive()),
  host = getOption("shiny.host", "127.0.0.1"),
  workerId = "",
  quiet = FALSE,
  display.mode = c("auto", "normal", "showcase"),
  test.mode = getOption("shiny.testmode", FALSE)
)
}
\description{
re-implementation of shiny::runApp
this version uses a more restricted range of ports
}
