% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_lib_paths.R
\name{.set_lib_paths}
\alias{.set_lib_paths}
\title{actually set R library paths
source: https://www.milesmcbain.com/posts/hacking-r-library-paths/}
\usage{
.set_lib_paths(lib_vec)
}
\description{
actually set R library paths
source: https://www.milesmcbain.com/posts/hacking-r-library-paths/
}
