% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_utils.R
\name{.setup_dev_dashboard}
\alias{.setup_dev_dashboard}
\title{example: learnr.dashboard:::.setup_dev_dashboard()}
\usage{
.setup_dev_dashboard(
  lib = "/home/siebrenf/git/edu/learnr.proto/renv/library/R-3.6/x86_64-pc-linux-gnu",
  pkg = "/home/siebrenf/git/edu/learnr.dashboard"
)
}
\description{
example: learnr.dashboard:::.setup_dev_dashboard()
}
