% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_utils.R
\name{.setup_learnr_dashboard}
\alias{.setup_learnr_dashboard}
\title{Install learnr.dashboard from project folder}
\usage{
.setup_learnr_dashboard(
  lib = NULL,
  pkg = "/home/siebrenf/git/edu/learnr.dashboard"
)
}
\description{
just a convenient wrapper
}
\examples{
learnr.dashboard:::.setup_learnr_dashboard()
}
\keyword{internal}
