% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_utils.R
\name{.setup_learnr_pkg}
\alias{.setup_learnr_pkg}
\title{Install learnr tutorials}
\usage{
.setup_learnr_pkg(tutorials = NULL, pkg = NULL, lib = NULL)
}
\arguments{
\item{tutorials}{List of tutorials to update. Will search the tutorials directory if unset.}

\item{pkg}{Tutorial package. Default: current directory}

\item{lib}{Library to install pkg. Default: library 1 (renv lib if activated)}
}
\description{
run from the base of the package project directory (or set the arguments)

1. git pull, update renv lib & install pkg in lib.

2. regenerate the HTML files for the listed tutorials

3. make all of lib accessible with "chmod -R 755"

4. make the path to lib accessible with "chmod 755"
}
\note{
uses system() with UNIX commands, may not be portable to Windows/Mac

if you get an "error in .validate_signal_args", try restarting the R session and run again.
}
\examples{
learnr.dashboard:::.setup_learnr_pkg("fg3")

learnr.dashboard:::.setup_learnr_pkg(c("fg1", "fg2"))
}
