% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_utils.R
\name{.setup_learnr_proto}
\alias{.setup_learnr_proto}
\title{Install learnr.proto from project folder}
\usage{
.setup_learnr_proto(lib = NULL, pkg = "/home/siebrenf/git/edu/learnr.proto")
}
\description{
just a convenient wrapper
}
\examples{
learnr.dashboard:::.setup_learnr_proto()
}
\keyword{internal}
