% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start_as_app.R
\name{.start_app}
\alias{.start_app}
\title{start tutorial from package as shiny markdown
example: learnr.dashboard:::.start_app("fg1")}
\usage{
.start_app(name = "test", package = "learnr.proto")
}
\description{
start tutorial from package as shiny markdown
example: learnr.dashboard:::.start_app("fg1")
}
