% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start_as_app.R
\name{.start_dev_app}
\alias{.start_dev_app}
\title{start dev tutorial as shiny markdown
works without reinstalling the package first
example: learnr.dashboard:::.start_dev_app("fg1")}
\usage{
.start_dev_app(
  name = "test",
  tutorial_dir = "/home/siebrenf/git/edu/learnr.proto/inst/tutorials"
)
}
\description{
start dev tutorial as shiny markdown
works without reinstalling the package first
example: learnr.dashboard:::.start_dev_app("fg1")
}
