% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start_as_tutorial.R
\name{end_background_tutorial}
\alias{end_background_tutorial}
\title{Stop all background tutorials}
\usage{
end_background_tutorial(...)
}
\arguments{
\item{...}{Ignored.}
}
\description{
Stop all learnr tutorials you are running in the background. Progress should be saved.
}
\details{
Uses system calls which may not be platform independent.
}
\examples{
learnr.dashboard::end_background_tutoria()

learnr.dashboard::end_background_tutoria("fg1")
}
