% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_renv_lib.R
\name{get_renv_lib}
\alias{get_renv_lib}
\title{Return path to the renv library}
\usage{
get_renv_lib(renv_lib = NULL)
}
\description{
Check if exactly one renv library is available, and if the library is usable.
Returns the path to the library.
}
\keyword{internal}
