% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_system_libs.R
\name{get_system_libs}
\alias{get_system_libs}
\title{Return path(s) to the system libraries}
\usage{
get_system_libs(system_libs = NULL)
}
\description{
Checks if the libraries are available and usable.
Returns the paths to the libraries.
}
\keyword{internal}
