% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_to_lib.R
\name{install_to_lib}
\alias{install_to_lib}
\title{install package from project folder to lib}
\usage{
install_to_lib(pkg = NULL, lib = NULL)
}
\description{
install package from project folder to lib
}
\examples{
learnr.dashboard:::install_to_lib("/home/siebrenf/git/edu/learnr.dashboard")
}
\keyword{internal}
