% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_lib_paths.R
\name{set_lib_paths}
\alias{set_lib_paths}
\title{actually set R library paths}
\usage{
set_lib_paths(lib_vec)
}
\arguments{
\item{lib_vec}{library path(s) to use}
}
\description{
R has no formal method to set library paths. Only Renv and this function seem to be able to do this.
}
\details{
source: https://www.milesmcbain.com/posts/hacking-r-library-paths/
}
\keyword{internal}
