% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_tutorial_libraries.R
\name{set_tutorial_libraries}
\alias{set_tutorial_libraries}
\title{Set libraries required to run learnr tutorials}
\usage{
set_tutorial_libraries(renv_lib = NULL, system_libs = NULL)
}
\description{
Assert that a renv library is available.
Sets the renv library and system libraries for the session.
}
\note{
Also preloads shiny and learnr, smoothing \code{start_background_tutorials()} by reducing the log.
}
\keyword{internal}
