% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start_as_tutorial.R
\name{start_tutorial}
\alias{start_tutorial}
\title{Start a learnr tutorial}
\usage{
start_tutorial(name, package = "learnr.proto")
}
\arguments{
\item{name}{Tutorial name. Available tutorials: "fg1", "fg2", "fg3", "fg4".}

\item{package}{Tutorial package. Default: learnr.proto (Functional Genomics)}
}
\description{
Start a learnr tutorial
}
\examples{
learnr.dashboard::start_tutorial("fg1")

learnr.dashboard::start_tutorial("fg2")
}
