% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_utils.R
\name{.start_dev_app}
\alias{.start_dev_app}
\title{start tutorial in development as shiny markdown}
\usage{
.start_dev_app(
  name = "test",
  tutorial_dir = "/home/siebrenf/git/edu/learnr.proto/inst/tutorials"
)
}
\description{
works without reinstalling the package first
}
\examples{
learnr.dashboard:::.start_dev_app("fg1")
}
\keyword{internal}
